Initialize rails app
Why these changes are being introduced:

We need a new rails 7 app.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-67

How this addresses that need:

This generates a new rails 7 app.

Side effects of this change:

It's unclear to me how many of the generated files we'll need, so
I've left it as is for now.