Add MIT Libraries theme gem
Why these changes are being introduced:

We want to style our UI in a way that is consistent with MIT
Libraries branding.

Relevant ticket(s):

Closes #16.

How this addresses that need:

This adds the mitlibraries-theme gem to the application.

Side effects of this change:

To confirm the changes, I created a basic search controller and
view. It doesn't do anything yet, except say congratulations in
German.