UI flow diagram
Why are these changes being introduced:

* to provide a visual representation of the application flow

Relevant ticket(s):

* rdi-71

How does this address that need:

* provides a mermaid-js flowchart in the readme

Document any side effects to this change:

As we make changes in the implementation, we'll need to make sure this
flow remains update or it will become misleading.

This is not a direct object relationship diagram, but a logical flow of
data through objects. It is possible to misinterpret it and thus
misunderstand how the application components actually communicate (which
this does not show).