Add importmap tags to head
Why these changes are being introduced:

This came about because of an uncaught syntax error in the browser
console regarding our import statements in application.js. We
suspected that this error indicated that we were not loading
Hotwire/Turbo properly.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-75

How this addresses that need:

This forks the head template from the theme gem in order to add the
JS importmap tags required for importmap to load our JS properly.

Side effects of this change:

* The error that flagged us to this problem persists, but it seems
innocuous as our JS still loads and Turbo frames work as expected.
* In addition to the `javascript_importmap_tags` method, the custom
template includes `csrf_meta_tags`, `csp_meta_tag`, and
`stylesheet_link_tag`, all of which are included in the default
Rails 7 template and seem harmless to add.

Co-authored-by: Jeremy Prevost <jprevost@mit.edu>