Speculative rebuild of timdex wrapper
This creates a parallel model of the Timdex class, using methods
rather than constants where possible. The parallel, TimdexCandy,
will work in the console - but I've been unable to get it working
from the controller.

The steps in the console are:

$ query = {'q' => 'data'}
$ wrapper = TimdexCandy.new
$ Parsed = wrapper.client.parse(TimdexCandy::SearchQueryText)
$ wrapper.search(Parsed, query)

The issue seems to be how to create the Parsed constant - which
needs the Client to exist for the parsing, but need to be stored
as a constant (because if you send a non-constant to the graphql
library you get an ugly error complaining about dynamic query
errors)

Maybe someone else, even future-me, will be able to resolve this.