Initial refinements to search results UI
Why these changes are being introduced:

This work was prompted by the initial set of UI iterations.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/RDI-99
* https://mitlibraries.atlassian.net/browse/RDI-138
* https://mitlibraries.atlassian.net/browse/RDI-144
* https://mitlibraries.atlassian.net/browse/RDI-145
* https://mitlibraries.atlassian.net/browse/RDI-146
* https://mitlibraries.atlassian.net/browse/RDI-147
* https://mitlibraries.atlassian.net/browse/RDI-149

How this addresses that need:

This proposes a path for the results UI design, focused on the
minimal information needed to identify whether a resource is useful.

Since we will have a full record view, these brief results include
only title, content type, publication date, contributors, and
full-text link. Unlike Bento, they do not include subjects, as I
couldn't discern a use case for this in the brief results. (More on
this in the ticket if you're curious.)

This makes some improvements to a11y, including highlighting
links and adding `:visited` classes. The colors of these are
subject to change; I'm specifically interested in the
reviewer's opinion of the green used for the 'View online' button;
this is borrowed from Bento, but I wonder if it has implications
for color-blind people.

These changes are very much a starting point. I anticipate
additional iteration on the design, and I welcome whatever changes
the reviewer recommends to this current state.

Side effects of this change:

* Contributor types have been removed, but I'm very open to adding
them back in, or possibly changing the number/type of contributors
shown.
* The 'View online' links don't actually go anywhere yet.
* There are some additional features that I think are needed, but
seemed out of scope of this spike. I created implementation tickets
for those that are listed under 'Relevant tickets' and linked from
the JIRA ticket.
* I still have many open questions on this approach, primarily
around the assumptions made re: affordances and metadata. See my
comment in the ticket for more info.