Add actionable links to ISBN fact panel
Why these changes are being introduced:

The current iteration of this fact panel does not provide
affordances to get the item with the identified ISBN.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-132

How this addresses that need:

This adds an a tag to the title with the Primo link resolver URL
for the given ISBN.

Side effects of this change:

* This does not include styling updates, as we've decided to
postpone that work.
* Author names have been added to the panel.
* I refactored the actual fetching and response parsing into a
separate method, so as to avoid repetition in the separate `fetch`
methods for ISBNs and authors.
* If we don't have the specific edition that the ISBN refers to,
but we do have other editions of the book, the link resolver isn't
sophisticated enough to indicate this. It might be useful in the
future to indicate in the fact panel whether other editions are
available, but we will likely want UXWS input on this.