Add highlighted search terms to results
Why these changes are being introduced:

The TIMDEX API now returns search term matches, so we should
display these in the UI.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/RDI-144
* https://mitlibraries.atlassian.net/browse/RDI-156

How this addresses that need:

This updates the schema and the TimdexSearch model to include the API
changes and adds a highlights partial to render matched terms.
The intent is to highlight term for fields that are not already in the
result view, which is achieved with a couple of helper methods.

It is also possible to highlight matched terms in fields that are
already shown in the results by swapping out the existing field with
the highlight fragment if present. Currently, this would affect
title, pub date, content type, and contributors. I chose not to do
this at this stage because we can't reliably assume that highlight
fragments are the same the original field, partly because they are
limited to 100 characters. We could change in this limit in OpenSearch,
but that would have other implications. This would be a good topic to check
in with UXWS when they can join us on the UI work.

Side effects of this change:

* Due to the changes to the schema, all cassettes that hit TIMDEX
API have been regenerated, and special conditions have been noted
on a couple of tests.
* Currently we are rendering _all_ matches, excluding fields
displayed in the result. Since a single field can have multiple
matches, this can result in a relatively long list of hits for, say,
the 'summary' field. This doesn't seem to be a major issue so far,
but we should check in with UXWS on it.