Add context about application
Why these changes are being introduced:

We want to provide users with information about the app. This
should be easily configurable, as we are configuring multiple experimental
interfaces under different URLs.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-213

How this addresses that need:

This adds an 'about' partial that renders the content of the
environment variable `ABOUT_APP`.

Side effects of this change:

* Storing context about the app in env is a bit messy, but it seems to
be the best approach here.
* A new platform (`x86_64-darwin-21`) was added to Gemfile.lock as a
result of running `bundle install`. Documenting here in case it causes
problems.