Fix name mismatch for fundingInformation
Why these changes are being introduced:

In the advanced search form, the name attribute for fundingInformation
is incorrectly "Funding information". This passes the wrong field name
to params and causes all fundingInformation searches originating from
the search form to fail.

Relevant ticket(s):

N/A.

How this addresses that need:

This updates the form partial to use the correct field name.

Side effects of this change:

My initial instinct was to write a regression test for this, but it
felt like opening a can of worms to test each input in the advanced
search form. I'd like the reviewer's input on this.