Experimental i18n implementation
Why these changes are being introduced:

We'd like to explore internationalization in our Rails applications
as a potential step toward making our platforms more accessible to
a global audience. TIMDEX UI is a good candidate for experimentation,
as it's still under active development.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/TIMX-63

How this addresses that need:

This is a first attempt at localizing English (US) in TIMDEX UI. As
it is a large changeset with significant architectural implications,
it should be considered a draft PR for team review, so we can see and
discuss what i18n might look like in our ecosystem.

Side effects of this change:

* In order to test the new controller method, an i18n test section
has been added to the style guide view. We don't want to inadvertently
mislead end users into thinking we've translated the app into a real
language, so we're testing with 'ewok' for now.
* The record/misc view hasn't been localized because we're not yet
sure what to do with the data. It's worth noting, however, that this
view uses a lot of helper method logic that may significantly
increase the complexity of localization.
* Several key architectural decisions are inferred here, including
the structure of the locale files, the way in which users switch
locales, and the extent to which we attempt to localize data from the
API. We will want to ADR these (and others) if we choose to move forward
with i18n.
* Our 'about this app' contextual information is no longer stored in ENV,
as this would preclude it from being localized.
* Support for right-to-left and bidirectional will require additional
considerations.