Display complete input values in advanced search form
Why these changes are being introduced:

When the advanced search form is submitted, the full search term
is not retained if it includes a space.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-237

How this addresses that need:

This wraps advanced search value attributes with quotations, so Rails
knows to retain the full param.

Side effects of this change:

None.