Delimit content types in a readable format
Why these changes are being introduced:

The record view does not join the content type field on a character,
so multiple content types are displayed in an unreadable format.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/RDI-242

How this addresses that need:

This joins the content type field on a semicolon to render it
the same as the result view.

Side effects of this change:

None.