Add Sentry for exception monitoring
Why these changes are being introduced:

We want to prepare for TIMDEX-UI as a potential production
application.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/TIMX-81

How this addresses that need:

This adds sentry-ruby and sentry-rails, including the required
initializer.

Side effects of this change:

* Configures Sentry environment. We haven't done this yet in our
Rails apps, but it seems like it will be especially useful for
TIMDEX and TIMDEX-UI, for which we have a lot of different
environments.
* Added SENTRY_ENV and SENTRY_DSN to Heroku config vars.