Explicitly specify Heroku stack
Why these changes are being introduced:

We do not currently declare the stack we want to use for PR builds.
This is fine as TIMDEX-UI is already using Heroku 22, but it may
be useful for future stack upgrades.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-180

How this addresses that need:

This adds an app.json file that declares the target stack and app
name.

Side effects of this change:

None.