Adds split gem with dashboard and redis
A/B testing will be handled by the Split gem. Because this PR will
include two different experiments as a demonstration, we also
add an initializer to enable that feature.

Along the way, we add a line to gitignore because local Redis
servers will create a file at project root. We also update the
project readme with a bit of information about running Redis
locally.