Add Split gem with dashboard route
Add reference to readme for how to run Redis

The split gem includes an instruction about how to run Redis (via homebrew), but I don't want to replicate those instructures - so this is mostly just a reminder to future-me about how to run it.

Ultimately this should be part of the devcontainer, or something similar?

Fixup