Implement a first experiment
This sets up a simple UI experiment, defining an alternative set of styles for the search form.

The search form starts in the _form.html.erb partial, with most users getting a "standard" class - the experimental group would get the "alt" class, which is rendered differently.

The experiment ends when a user reaches a search results page (i.e, they submitted the form and thus got results).

Curious folks can force themselves into one or the other pool by using a querystring such as

?ab_test[ui_colors]=alt
or
?ab_test[ui_colors]=standard