Second experiment
This tries implementing a second (concurrent) experiment, which would rely on knowing how often a user clicks an outbound link.

That means we need an interstitial redirect in order to conclude the experiment, which happens in the new outbound path.

PLEASE NOTE: This isn't working yet, because the redirect is getting caught by Turbo, it seems, and violating CORS policies. As a result, we're just logging the outbound URLs to the console.