Update rails to 7.1.2; adds note to Gemfile
The Rails update is pretty straightforward. One of our other gems is
pinned to a specific version due to an upstream issue. That was linked
in the commit message when it happened, but I'm adding it as a comment
in the Gemfile for easier follow-up down the road. Hopefully it doesn't
make a muddle of things.