Document decision to use ADRs
Why these changes are being introduced:

We are about to make major architectural decisions about this
application, and we need a framework for recording them.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-123

How this addresses that need:

This adds an initial architectural decision record (ADR) to confirm
that we will use ADRs in this repo.

Side effects of this change:

Future significant architectural decisions should be documented
as ADRs.