Document initial architectural decisions related to GDT
Why these changes are being introduced:

We plan to develop the GIS Data Tool UI as an extension of
TIMDEX UI. We need to do so without disrupting the existing
TIMDEX UI applications, while also planning for a potential
future state in which those applications may include geospatial
data.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-123

How this addresses that need:

This adds an ADR to confirm our initial plan for the GDT UI.

Side effects of this change:

* We will be adding the `flipflop` gem in a future PR to manage
a `gdt` feature flag.
* We will update the theme gem to add app names to headers.
* One decision in the ADR is contingent upon a decision we've yet
to make, regarding whether we will want to display GIS results
in non-GDT TIMDEX UI apps. I've phrased that as more of a potential
decision that will likely require a subsequent ADR if we choose to
go that route.