Update advanced search language
Why these changes are being introduced:

UXWS has requested that he simplify the language for the
advanced search dropdown.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-126

How this addresses that need:

This updates the language as requested: "Advanced search" when the
dropdown is closed, and "Close advanced search" when it's open.

Side effects of this change:

Though it emerged from the GDT UI design work, this is a general
enhancement that will apply to all TIMDEX UI apps.