Add GDT feature flag
Why these changes are being introduced:

We need a way to toggle GDT features, so we can manage non-GDT
TIMDEX UI apps alongside it.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/GDT-157

How this addresses that need:

This adds the `flipflop` gem and a GDT feature.

Side effects of this change:

* This is likely the first time we've run db migrations since
updating to Rails 7.1. Running migrations locally now generates
`*.sqlite3-shm` and `-wal` files. I've updated the gitignore
to ignore these (and all sqlite files).
* The `flipflop` gem doesn't appear to be very actively maintained
these days. We'll want to keep an eye on that and, if it gets
abandoned, consider switching to something like `flipper`.