Redesign filter sidebar
Why these changes are being introduced:

The filter sidebar can become unwieldy, as all available filters
are shown at all times.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-127
* https://mitlibraries.atlassian.net/browse/GDT-173
* https://mitlibraries.atlassian.net/browse/GDT-174

How this addresses that need:

This restyles the filter sidebar as dropdown menus. In so doing,
it also changes some behavior in the filter logic:

* Number of results is listed in the filter sidebar header.
* Clicking an applied filter removes that filter (this replaces the
previous removal mechanism, which removed all applied filters).
* Applied filters are highlighted with an `x` symbol, to indicate
that it can be removed on click. (Previously, the 'applied' filter
was added, but we hadn't done anything with it.)
* The first filter dropdown in the sidebar is always open, even if
no filters are applied.

Not addressed in this commit:

* Darcy asked that we persist menu state between page loads, such
that an open menu always remains open unless a user closes it
manually (e.g., if all filters have been removed in that category).
This proved relatively complex, so I opened GDT-173 to address it.
* There are some visual differences in how the sidebar appears
in smaller viewports. This is covered in GDT-174.

Side effects of this change:

* Some tests have been updated to reflect the new logic.
* One potential a11y issue with filter categories is having to
navigate through a bunch of filter opens before getting to the
next category. We might want to add skip-content links if this
feels onerous.
* The JS for the dropdown menus is so minimal that I included it
in the view as a `script` tag.
* The filter container is now above the results in the DOM.