Update theme gem and document PLATFORM_NAME variable
Why these changes are being introduced:

The GDT project added a feature to the theme gem to conditionally
include the platform name in the header if the `PLATFORM_NAME`
environment variable is supplied.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/GDT-124

How this addresses that need:

This updates the theme gem to the v1.2 tag, which includes the new
feature, and documents the PLATFORM_NAME variable.

Side effects of this change:

None. The app will continue to use the standard header if
PLATFORM_NAME is not supplied.