Redesign local nav for GDT
Why these changes are being introduced:

The Geodata app requires links to MIT GIS resources. It also does
not require the title of the app in the local header, as GDT-124
adds that to the main header.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-125
* https://mitlibraries.atlassian.net/browse/GDT-124

How this addresses that need:

This restyles the local nav as specified, provided the GDT feature
is enabled.

Side effects of this change:

Most likely, we will want to remove the app name from the local
header for all TIMDEX apps, as that will be included in the main
header for all Rails apps from now on. However, it's out of scope
of this ticket (and project) to make that decision.