Redesign local nav for GDT
Why these changes are being introduced:

The Geodata app requires links to GIS resources at MIT in the local
nav. It also no longer requires the app name in the local nav, as
GDT-125 moved this to the main header.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-125
* https://mitlibraries.atlassian.net/browse/GDT-124

How this addresses that need:

This restyles the local nav as specified, provided the GDT feature
is enabled. It also makes the title in the nav conditional on whether
the `PLATFORM_NAME` environment variable is present. (This toggles the
name in the header.)

Side effects of this change:

None.