Reorganize rules in timdexui.scss
Why these changes are being introduced:

There are some CSS rules in `timdexui.scss` with a note that
they should be moved out of a file if there's a better place to
put them. Since then, we've added several partials that might make
logical sense to contain to these rules.

Relevant ticket(s):

N/A.

How this addresses that need:

This makes the following changes to `timdexui.scss`:

* The `#advanced-search-panel legend` rule moves to the `panels`
partial. (Though, it would also make sense in the `search` partial.)
* The `.top-space` rule moves to the `results` partial, as it is only
used for the filter container spacing.
* The `.top-attached` and `.top-line` rules are removed. Currently,
`.top-line` is only used in the style guide layout, and
`.top-attached` appears not to be used anywhere.

Side effects of this change:

The style guide layout will no longer have a top border.