Add 'Ask GIS' panel
Why these changes are being introduced:

We want to provide users multiple paths to request help from the
GIS team.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/GDT-129

How this addresses that need:

This adds an 'Ask GIS' link to all views except the index:

* On the results page with reuslts, in the filter sidebar below
the filters.
* On the results page with no results, in a sidebar to the right.
* On the full record page, in the sidebar below fact panels.

There is a minor difference in placement for the results view in
smaller viewports, which is accommodated with media queries.

Side effects of this change:

The new `ask` partial contains both the logic for the existing
'Ask Us' link and the 'Ask GIS' link. In order to minimize the
complexity, both use the same markup. This removes the descriptive
paragraph for 'Ask Us' and changes the header from 'Ask Us' to
'Need help?', both of which seem like usability improvements.