Add search summary panel
Why these changes are being introduced:

It would be helpful if users could see what filters they've
applied to a search directly below the search bar, which would
also provide another path to remove filters.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/GDT-142

How this addresses that need:

This provides the requested panel. We had been calling this a
"filter removal" panel, but I ended up naming it "search summary"
to avoid any confusion with the filter sidebar.

Side effects of this change:

The "You searched for" list has been removed, as this feature
replaces it.