Add site title partial
Why these changes are being introduced:

The GeoData visual design includes a heading at the top of each
page, along with some descriptive text.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/GDT-182

How this addresses that need:

This adds the requested element as a partial, with conditional
copy for GDT versus non-GDT applications. It also moves the label text
to the `title` attribute, as the label is is now visually redundant but
may still be useful to screen reader users.

Side effects of this change:

* The input label before form submit is now the same as the placeholder
text. This feels fine since screen readers might not read placeholder
text, but they do consistently rely on labels.
* The copy for non-GDT titles did not come from UXWS and is subject
to change.
* The conditional title-in-nav has been removed from `site_nav`.
This seems superfluous since we now have h1 tags.