Make default progress indicator more prominent
** Why are these changes being introduced:

* Our application may already be showing a progress indicator that comes
  with Rails, but the default rendering is very subtle (only three px
  high). We want to make this more prominent.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/gdt-143

** How does this address that need:

This makes two changes:

* Sets the threshold for showing the progress to bar to 200 milliseconds
  instead of the default 500. This will result in users seeing the
  progress bar more quickly while waiting for a slightly slow page load,
  which should give them greater assurance that the application is
  working on a response.

* Once shown, the progress bar is now 75 pixels tall, rather than 3.
  This is the height of the black navbar at the top of the page, which
  should make the progress bar much harder to miss when it appears.

** Document any side effects to this change:

This isn't a side effect, but an issue I've become aware of which I
can't change at the moment: the default progress bar which Turbo uses is
not particularly accessible to screen readers, which means that the
"visual indicator of progress happening" is exactly that - purely
visual. The screen reader experience while the indicator moves is
completely silent.

This is not a new behavior with this change, so it isn't a regression -
but it is an example an improvement for one audience not being extended
to other audiences. I'm not sure whether this should be enough to block
this PR from merging.

The level of effort to make this solution accessible for all audiences
would be something like these options:

* Write some javascript to augment the default markup with more
  accessible markup (aria roles, etc)
* Contribute upstream to turbo-rails to replace the div being used by
  Turbo with more accessible markup
* Look for an existing gem that has already augmented the default markup
  (I've done some brief searching for such a gem, but nothing
  immediately jumped out at me)