Add all currently-available aggregations
** Why are these changes being introduced:

This app implements only two filter / aggregation categories: content
type and source. However, others are now supported by the API:
Contributors, format, languages, literary form, and subjects. Others are
coming soon, but for now these are what is available.

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/gdt-128

** How does this address that need:

This updates our search model, and filter helper, to add support for the
new categories.

** Document any side effects to this change:

Because all queries now include this request for more aggregations, we
need to regenerate a number of the test cassettes.