Update schema definition after recent API changes
** Why are these changes being introduced:

The TIMDEX API has made some recent changes to its schema, which need to
be reflected in this UI app as well.

** Relevant ticket(s):

n/a (Should there be a ticket for this?)

** How does this address that need:

This runs the console command in the UI app to update schema.json,

** Document any side effects to this change:

Some in-progress branches which have incorporated these changes on
their own will need to be tweaked to remove the diff from their scopes.