Allow applications to overide filter name via env
** Why are these changes being introduced:

* The nice_labels method in the filter helper defines static names for
  each possible filter category, but the GDT project makes it likely
  that these names may need to vary by application due to the nature of
  the records available (i.e. "content type" versus "data type").

** Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/gdt-128

** How does this address that need:

* This adds support for loading the filter name from ENV, using a set of
  variables listed in the README.

** Document any side effects to this change:

This is not a side effect, but a point of hesitation. I'm unsure if we
need tests for _each_ ENV override in nice_labels. Right now I've
written one to confirm that "Content type" is the default, and one to
confirm that it can be overriden - for that category only. It feels
like a bit of overkill to test each of the seven categories (and then
need to write more tests as new categories are added to the API) - but
this leaves open the possibility that the overrides - which are not
likely to be heavily used - might break in a way that we do not realize
initially.

I take some solace in the fact that the defaults we provide are "okay",
and that there will be stakeholders watching for the overridden values -
so any future problem should be caught relatively quickly - but I'm open
to contrasting perspectives here.