Address 'Ask Us' feedback from UXWS
Why these changes are being introduced:

Darcy suggested changing the color, spacing, and size of the chat
icon.

Relevant ticket(s):

* [GDT-129](https://mitlibraries.atlassian.net/browse/GDT-129)

How this addresses that need:

This applies the suggested changes.

Side effects of this change:

I'd originally intended to move the `view-md` and `view-lg` rules
to the `layouts` partial, but opted against it since those are
only used for this element. I'd be happy to make that change, though,
if the reviewer feels otherwise.