Add GDT access links
Why these changes are being introduced:

The GDT UI requires various access links to appear in the
record sidebar. The text and hrefs of these depend on the data
provider and access restrictions.

Relevant ticket(s):

* [GDT-141](https://mitlibraries.atlassian.net/browse/GDT-141)
* [GDT-175](https://mitlibraries.atlassian.net/browse/GDT-175)
* [GDT-176](https://mitlibraries.atlassian.net/browse/GDT-176)

How this addresses that need:

This applies the requested changes:

* A 'Download geodata files' link appears for MIT-provided
records that are OA.
* A 'Download geodata files' link with an auth notice appears
for MIT-provided records that require auth.
* A 'View `[provider]` record' link appears for non-MIT records,
where `[provider]` is the institution providing access to the
data.

Non-MIT records always use the `kind=Website` link, whereas MIT
records always use the `kind=Download` and `text=Data` link.

GDT-175 is *not* implemented in this commit, as we have not yet
loaded GIS records from Alma.

Side effects of this change:

* The schema has been updated, and some unrelated cassettes have
been regenerated as a result.
* The `provider` field has been added to the `Record` query
object.
* A `record` SCSS partial has been added.
* The controller tests rely on specific data returned by the API
call. We've generally avoided this testing approach as it can be
brittle, but this seemed like the best way to test the
conditional logic in the new `access_button` partial. We do have
tests covering the helper methods, so we could consider removing
these tests for now. Either way, we should evaluate E2E testing
for this repo (and likely others).