Make new aggregations searchable
Why these changes are being introduced:

We added some more aggregations as part of the ongoing work, but did
not add them to the query string or corresponding models.

Relevant ticket(s):

* (GDT-128)[https://mitlibraries.atlassian.net/browse/GDT-128]

How this addresses that need:

This adds the new aggregations to the Enhancer and QueryBuilder models,
and adds them to the query in the TimdexSearch model.

Side effects of this change:

* VCR cassettes have been regenerated.
* In the filter partial, `to_sym` calls on the filter category have
been removed. These are no longer needed as all filter categories
are cast as symbols in the search controller.
* This is another instance of something that has come up recently:
the need to develop an E2E testing strategy for our Rails apps,
particularly those that are API consumers. We can test this using
our standard VCR process, but that is likely to break when
cassettes are regenerated and the search results change.