Update result view for GDT
Why these changes are being introduced:

The GeoData app requires specific metadata in the results view.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-130

How this addresses that need:

This adds a separate result view for geo records that includes
the specifed metadata fields. It also applies some general styling
updates, most of which involve spacing and font size/weight. The
most significant CSS change is that values that may be long
(authors, higlights, summary) are now truncated.

Side effects of this change:

* The date parsing logic catches a lot of edge cases, but not all
of them. It also puts a fair amount of faith in the source data to
have reasonably formatted dates. Finally, it makes the assumption
that the date we want is coverage, falling back to issued. This is
subject to change.
* I noticed an indentation issue in the search controller test, so
some lines that appear to have changed have just been re-indented.
New tests begin at line 1036.
* Certain fields have been added to support these changes, and
cassettes have been regenerated as a result.